refactor: login flow
